+++ weave/trace/api.py
@@ -9,8 +9,7 @@ from typing import Any, Callable, Iterator, Optional, Union
 # TODO: type_serializers is imported here to trigger registration of the image serializer.
 # There is probably a better place for this, but including here for now to get the fix in.
 from weave import type_serializers  # noqa: F401
-from weave.legacy.weave import urls
-from weave.trace import util
+from weave.trace import urls, util
 from weave.trace.call_context import get_current_call, require_current_call
 from weave.trace.client_context import weave_client as weave_client_context
